requirements.txt
aioredis==1.3.1
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.2.2
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.6
daphne==3.0.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.2.4
django-allauth==0.44.0
django-cors-headers==3.7.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-stubs==1.8.0
django-stubs-ext==0.2.0
djangorestframework==3.12.4
djangorestframework-stubs==1.4.0
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
msgpack==1.0.2
mypy==0.902
mypy-extensions==0.4.3
oauthlib==3.1.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==2.0.1
pyOpenSSL==20.0.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==20.3.0
txaio==21.2.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.2
whitenoise==5.2.0
zope.interface==5.2.0
